Bot v2.0
Bot Is based on moving avergaes and trendline similar to the v1.0 but now also includes ICT Concepts such as FVG and order-blocks for confirmation and trade entry.
Trades are done on the 1 minute timeframe with 5 and 15 minutes timeframe serving as the analysis timeframe. 

Bot has an fixed risk to reward ratio that it follows for every trades that is being taken by the Mt-bot.